Refactored usage of setting the title as edited or not
window now sets the title by default. Two new methods implemented:
setTitleEdited and getTitleState.

setTitleEdited takes a boolean as an argument. Passing True sets the
state as edited and changes the title to *GPytage. Passing False sets
the state back to normal and reverts the change to the title.

getTitleState returns a boolean of the current title state.