Massive refactoring of datastore to use the new FolderObj and PackageFileObj objects
folderModel, the gtk.TreeStore, now only contains two colums: A String
and an Object column. The columns are referred to by constants as
F_NAME and F_REF respectively.

initData() constructs the Folder and PackageFile objects with their
respective references to the filesystem.

initTreeModel() populates the TreeModel with data gathered from
initData(). This only populates the folderModel, as each PackageFile
object creates their own internal gtk.ListStore to keep track of each
files contents.

clearData() has been implemented to allow for easy clearing of all
current data in the Models. It effectively clears the folderModel and
deletes all Folder and PackageFile objects collected in the TLFolders
and TLFiles list.