Inserting a package now functional via right click
helper.py:
fileEdited() now sets the PackageFile objects edited state to True.

getMultiSelection() implemented for returning TreeRowReferences of the
currently selected rows in the rightpanel.

getCurrentFile() implemented for when a row is being inserted into a
blank file, or the user has not clicked on a preexisting row. It
returns the PackageFileObj and the leftpanels model.

panelfunction.py:
Removed old mselected() method.

rightpanel.py:
Implemented the new insertRow() method. This allows a user to insert a
row below the rightclicked row or append the row to the next available
row if the file happens to be blank or the user clicked on a
non-existent row.